12/5 Removing the leaves from testing and adding snow textures to go with the new particle weather

+++ src/main/java/com/otterly76/ott/particle/StreakParticle.java
@@ -56,7 +56,7 @@ public class StreakParticle extends WeatherParticle {
             }
 
             this.gravity = 0.0F;
-            this.yd = (double)0.0F;
+            this.yd = 0.0F;
             this.shouldFadeOut = true;
         }
